allow deleting disabled users
even if it is spacewalk/satellite admin

+++ java/code/src/com/redhat/rhn/manager/user/UserManager.java
@@ -492,7 +492,7 @@ public class UserManager extends BaseManager {
 
         // Do not allow deletion of the last Satellite Administrator:
         User toDelete = UserFactory.lookupById(loggedInUser, targetUid);
-        if (toDelete.hasRole(RoleFactory.SAT_ADMIN)) {
+        if (!toDelete.isDisabled() && toDelete.hasRole(RoleFactory.SAT_ADMIN)) {
             if (SatManager.getActiveSatAdmins().size() == 1) {
                 log.warn("Cannot delete the last Satellite Administrator");
                 throw new DeleteSatAdminException(toDelete);
